Extract npm-engine as a lib
