+++ buildpacks/nodejs-npm-engine/src/node.rs
@@ -3,12 +3,12 @@ use libcnb::Env;
 use std::process::Command;
 
 #[derive(Debug)]
-pub(crate) enum VersionError {
+pub enum VersionError {
     Command(CmdError),
     Parse(String, heroku_nodejs_utils::vrs::VersionError),
 }
 
-pub(crate) struct Version<'a> {
+pub struct Version<'a> {
     pub(crate) env: &'a Env,
 }
