+++ buildpacks/nodejs-npm-engine/src/npm.rs
@@ -3,7 +3,7 @@ use libcnb::Env;
 use std::process::Command;
 
 #[derive(Debug)]
-pub(crate) enum VersionError {
+pub enum VersionError {
     Command(CmdError),
     Parse(String, heroku_nodejs_utils::vrs::VersionError),
 }
